[Chromium] Merge BrowserInitializer into RuntimeImpl
This class was initially refactored from RuntimeImpl. It does however
not make much sense to have it in a separate class. It creates unnecessary
differences with the Gecko backend for no reason.
