+++ app/src/common/chromium/com/igalia/wolvic/browser/api/impl/RuntimeImpl.java
@@ -3,6 +3,7 @@ package com.igalia.wolvic.browser.api.impl;
 import android.app.Service;
 import android.content.Context;
 import android.content.res.Configuration;
+import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
 
@@ -13,10 +14,12 @@ import com.igalia.wolvic.browser.api.WResult;
 import com.igalia.wolvic.browser.api.WRuntime;
 import com.igalia.wolvic.browser.api.WRuntimeSettings;
 import com.igalia.wolvic.browser.api.WWebExtensionController;
+import com.igalia.wolvic.utils.SystemUtils;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
 
 import kotlin.Lazy;
 
@@ -24,20 +27,36 @@ import mozilla.components.concept.fetch.Client;
 import mozilla.components.concept.storage.LoginsStorage;
 import mozilla.components.lib.fetch.httpurlconnection.HttpURLConnectionClient;
 
+import org.chromium.base.CommandLine;
+import org.chromium.base.ContextUtils;
+import org.chromium.base.PathUtils;
+import org.chromium.base.library_loader.LibraryLoader;
+import org.chromium.base.library_loader.LibraryProcessType;
+import org.chromium.content_public.browser.BrowserStartupController;
+import org.chromium.content_public.browser.DeviceUtils;
+import org.chromium.ui.base.ResourceBundle;
 import org.chromium.wolvic.VRManager;
 
 public class RuntimeImpl implements WRuntime {
-    private BrowserInitializer mBrowserInitializer;
+    static String LOGTAG = SystemUtils.createLogtag(RuntimeImpl.class);
     private Context mContext;
     private WRuntimeSettings mRuntimeSettings;
     private WebExtensionControllerImpl mWebExtensionController;
     private ViewGroup mContainerView;
+    private boolean mIsReady = false;
+    private ArrayList<RuntimeImpl.Callback> mCallbacks;
+    private static final String PRIVATE_DATA_DIRECTORY_SUFFIX = "content_shell";
+
+    interface Callback {
+        default void onReady() {}
+    }
 
     public RuntimeImpl(@NonNull Context ctx, @NonNull WRuntimeSettings settings) {
         mContext = ctx;
         mRuntimeSettings = settings;
         mWebExtensionController = new WebExtensionControllerImpl();
-        mBrowserInitializer = new BrowserInitializer(ctx);
+        mCallbacks = new ArrayList<>();
+        initBrowserProcess(mContext);
     }
 
     @NonNull
@@ -125,8 +144,59 @@ public class RuntimeImpl implements WRuntime {
         return mContainerView;
     }
 
-    @NonNull
-    public BrowserInitializer getBrowserInitializer() {
-        return mBrowserInitializer;
+    public void registerCallback(@NonNull RuntimeImpl.Callback callback) {
+        if (mIsReady) {
+            callback.onReady();
+        } else {
+            mCallbacks.add(callback);
+        }
     }
+
+    public void unregisterCallback(@NonNull RuntimeImpl.Callback callback) {
+        // It is possible that `callback` is already removed on the browser process startup
+        // callback.
+        mCallbacks.remove(callback);
+    }
+
+    private void initBrowserProcess(Context context) {
+        assert isBrowserProcess() == true;
+
+        ContextUtils.initApplicationContext(context);
+        ResourceBundle.setNoAvailableLocalePaks();
+        // Native libraries for child processes are loaded in its implementations in content.
+        LibraryLoader.getInstance().setLibraryProcessType(LibraryProcessType.PROCESS_BROWSER);
+
+        PathUtils.setPrivateDataDirectorySuffix(PRIVATE_DATA_DIRECTORY_SUFFIX);
+
+        CommandLine.init(new String[] {});
+        DeviceUtils.addDeviceSpecificUserAgentSwitch();
+        LibraryLoader.getInstance().ensureInitialized();
+
+        BrowserStartupController.getInstance().startBrowserProcessesAsync(
+                LibraryProcessType.PROCESS_BROWSER, true /* startGpuProcess */, false /* startMinimalBrowser */,
+                new BrowserStartupController.StartupCallback() {
+                    @Override
+                    public void onSuccess() {
+                        Log.i(LOGTAG, "The browser process started!");
+                        mIsReady = true;
+                        mCallbacks.forEach(callback -> {
+                            callback.onReady();
+                        });
+                        mCallbacks.clear();
+                    }
+
+                    @Override
+                    public void onFailure() {
+                        Log.e(LOGTAG, "Failed to start the browser process");
+                        // Clear callbacks on failure. This is needed as long as we don't retry to
+                        // start the browser process.
+                        mCallbacks.clear();
+                    }
+                });
+    }
+
+    private static boolean isBrowserProcess() {
+        return !ContextUtils.getProcessName().contains(":");
+    }
+
 }
